extracting context menu into component
