+++ visualization/app/codeCharta/ui/ui.ts
@@ -18,6 +18,7 @@ import "./experimentalSettingsPanel/experimentalSettingsPanel";
 import "./regexFilter/regexFilter";
 import "./resetSettingsButton/resetSettingsButton";
 import "./dialog/dialog";
+import "./nodeContextMenu/nodeContextMenu";
 
 angular.module(
     "app.codeCharta.ui",
@@ -37,6 +38,7 @@ angular.module(
         "app.codeCharta.ui.rangeSlider",
         "app.codeCharta.ui.resetSettingsButton",
         "app.codeCharta.ui.dialog",
+        "app.codeCharta.ui.nodeContextMenu",
         "app.codeCharta.ui.regexFilter"
     ]);
