+++ visualization/app/codecharta/ui/nodeContextMenu/nodeContextMenu.component.scss
@@ -1,0 +1,4 @@
+color-settings-panel-component {
+
+}
+
